service-workers/service-worker/fetch-request-css-base-url.https.html

<!DOCTYPE html>
<title>Service Worker: CSS's base URL must be the request URL even when fetched from other URL</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js?pipe=sub"></script>
<script src="resources/recorder.js"></script>
<script>
async_test(function(t) {
    var SCOPE = 'resources/fetch-request-css-base-url-iframe.html';
    var SCRIPT = 'resources/fetch-request-css-base-url-worker.js';
    var worker, frame;

    return service_worker_unregister_and_register(t, SCRIPT, SCOPE)
      .then(function(registration) {
          worker = registration.installing;
          return wait_for_state(t, worker, 'activated')
			.then(function() {
			    return recorder.clear(registration.active);
			  });
        })
      .then(function() { return with_iframe(SCOPE); })
      .then(function(f) {
		  frame = f;
		  console.log(f.contentDocument.getElementsByTagName('link')[0]);
		  return recorder.read(worker);
		})
	  .then(function(results) {
          var base = get_host_info()['HTTPS_ORIGIN'] + base_path();

		  assert_equals(results.length, 1);
          assert_equals(
            results[0].url,
            base + 'resources/dummy.png',
            'The base URL while loading the images referred from CSS ' +
            'must be the request URL of CSS.');
          assert_equals(
            results[0].referrer,
            base + 'resources/fetch-request-css-base-url-style.css',
            'While loading the image defined in CSS the referrer must ' +
            'be the request URL of CSS.');

          //frame.remove();

          return service_worker_unregister_and_done(t, SCOPE);
        })
      .catch(unreached_rejection(t));
  }, 'CSS\'s base URL must be the request URL even when fetched from other URL.');
</script>
